feat(publications): created the interfaces layer and container file

+++ src/modules/publication/infrastructure/repositories/mongo/publicationMongo.repository.ts
@@ -1,8 +1,8 @@
+import mongoose from "mongoose";
+
 import { Publication } from "@src/modules/publication/domain/entities/publication";
 import { PublicationRepository } from "@src/modules/publication/domain/repositories/publication.repository";
 import { PublicationModel } from "../../models/publication.model";
-import { IUserDocument } from "@src/shared/models/user.model";
-import mongoose from "mongoose";
 import { AppError } from "@src/shared/utils";
 import { ERROR_MESSAGES, HTTP_CODES } from "@src/shared/constants";
 
@@ -10,15 +10,13 @@ export class PublicationMongoRepository implements PublicationRepository {
   async getAllByUser(
     userAuthenticated: string,
     page = 1,
+    pageSize = 10,
   ): Promise<Publication[] | []> {
-    const PAGE_SIZE = 10;
-    const skip = (page - 1) * PAGE_SIZE;
+    const skip = (page - 1) * pageSize;
     const docs = await PublicationModel.find({ userId: userAuthenticated })
-      .populate<{ userId: IUserDocument }>({
-        path: "userId",
-      })
+
       .skip(skip)
-      .limit(PAGE_SIZE)
+      .limit(pageSize)
       .lean();
 
     if (!docs) {
@@ -37,6 +35,13 @@ export class PublicationMongoRepository implements PublicationRepository {
   }
 
   async create(publication: Publication): Promise<Publication> {
+    if (!publication.getUserId()) {
+      throw new AppError(
+        ERROR_MESSAGES.USER_ID_IS_REQUIRED,
+        HTTP_CODES.BAD_REQUEST,
+      );
+    }
+
     const result = await PublicationModel.create({
       userId: new mongoose.Types.ObjectId(publication.getUserId()),
       text: publication.getText(),
@@ -50,29 +55,15 @@ export class PublicationMongoRepository implements PublicationRepository {
       result._id.toString(),
     );
   }
-  async update(publication: Publication): Promise<Publication | null> {
-    const doc = await PublicationModel.findById(publication.getId()).lean();
 
-    if (!doc) {
-      throw new AppError(
-        ERROR_MESSAGES.PUBLICATION_NOT_FOUND,
-        HTTP_CODES.NOT_FOUND,
-      );
-    }
-
-    if (doc.userId.toString() !== publication.getUserId()) {
-      throw new AppError(
-        ERROR_MESSAGES.FORBIDDEN_RESOURCE_MODIFICATION,
-        HTTP_CODES.FORBIDDEN,
-      );
-    }
-
+  async update(
+    publicationId: string,
+    text: string,
+  ): Promise<Publication | null> {
     const result = await PublicationModel.findByIdAndUpdate(
-      publication.getId(),
+      publicationId,
       {
-        userId: publication.getUserId(),
-        text: publication.getText(),
-        attachments: publication.getAttachments(),
+        text,
       },
       { new: true },
     );
@@ -80,6 +71,7 @@ export class PublicationMongoRepository implements PublicationRepository {
     if (!result) {
       return null;
     }
+
     return new Publication(
       result.userId.toString(),
       result.text,
@@ -87,7 +79,16 @@ export class PublicationMongoRepository implements PublicationRepository {
       result._id.toString(),
     );
   }
+
   async delete(publicationId: string): Promise<void> {
+    const doc = await PublicationModel.findById(publicationId);
+
+    if (!doc) {
+      throw new AppError(
+        ERROR_MESSAGES.PUBLICATION_NOT_FOUND,
+        HTTP_CODES.NOT_FOUND,
+      );
+    }
     await PublicationModel.findByIdAndDelete(publicationId);
   }
 }
